Add files via upload
This report is an observational data study that aimed to answer if the biannual time change had a causal effect on the number of traffic collisions in the city of Los Angeles. The data used in this report was both acquired from the LA Data Portal website (collision data) and generated by myself using the aid of a time and date website (time change date data).