Add files via upload
Added files for Toronto Traffic During Pandemic project. This includes the Rmd file, the PDF report that was knitted from the Rmd file, and the cleaned data file that only includes the most relevant data for the report, and can be shared due to it being open data. (The source of the original data is referenced in the report itself.)